refactored types.js into examTypes
